[ty] Restore independent loop-header type inference
